# docs/dz8/swagger.yaml
openapi: 3.0.0
info:
  title: Микросервисная платформа электронной коммерции
  description: |
    API для микросервисной платформы электронной коммерции, включающей сервисы заказов, 
    оплаты, доставки, склада, биллинга и уведомлений
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Сервис заказов
  - url: http://localhost:8081
    description: Сервис биллинга
  - url: http://localhost:8082
    description: Сервис уведомлений
  - url: http://localhost:8083
    description: Сервис платежей
  - url: http://localhost:8084
    description: Сервис склада
  - url: http://localhost:8085
    description: Сервис доставки
...
paths:
...
  /api/v1/warehouse/{id}:
    get:
      tags:
        - Warehouse
      summary: Получение товара по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о товаре
        '404':
          description: Товар не найден
  /api/v1/warehouse/product/{product_id}:
    get:
      tags:
        - Warehouse
      summary: Получение товара по ID продукта
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о товаре
        '404':
          description: Товар не найден
  /api/v1/warehouse:
    get:
      tags:
        - Warehouse
      summary: Получение списка всех товаров
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список товаров
  /api/v1/warehouse/check:
    post:
      tags:
        - Warehouse
      summary: Проверка наличия товаров
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
  /api/v1/warehouse/reserve:
    post:
      tags:
        - Warehouse
      summary: Резервирование товаров
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                user_id:
                  type: integer
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      quantity:
                        type: integer
      responses:
        '200':
          description: Товары зарезервированы
        '400':
          description: Ошибка в запросе или недостаточно товаров
  /api/v1/warehouse/release:
    post:
      tags:
        - Warehouse
      summary: Освобождение резервации товаров
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                user_id:
                  type: integer
      responses:
        '200':
          description: Резервация отменена
        '400':
          description: Ошибка в запросе
  /api/v1/warehouse/confirm:
    post:
      tags:
        - Warehouse
      summary: Подтверждение резервации товаров
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: integer
                user_id:
                  type: integer
      responses:
        '200':
          description: Резервация подтверждена
        '400':
          description: Ошибка в запросе
  /api/v1/warehouse/order/{order_id}:
    get:
      tags:
        - Warehouse
      summary: Получение резерваций для заказа
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список резерваций
        '404':
          description: Резервации не найдены

  # Сервис склада - Внутренние API
  /internal/warehouse/reserve:
    post:
      tags:
        - Warehouse
        - Internal
      summary: Внутреннее резервирование товаров
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveWarehouseRequest'
      responses:
        '200':
          description: Товары зарезервированы
        '400':
          description: Ошибка в запросе или недостаточно товаров
        '500':
          description: Внутренняя ошибка сервера
  /internal/warehouse/release:
    post:
      tags:
        - Warehouse
        - Internal
      summary: Внутреннее освобождение резервации товаров
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseWarehouseRequest'
      responses:
        '200':
          description: Резервация отменена
        '400':
          description: Ошибка в запросе
        '500':
          description: Внутренняя ошибка сервера
  /internal/warehouse/confirm:
    post:
      tags:
        - Warehouse
        - Internal
      summary: Внутреннее подтверждение резервации товаров
      description: Эндпоинт для вызова другими сервисами, не требует стандартной аутентификации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmWarehouseRequest'
      responses:
        '200':
          description: Резервация подтверждена
        '400':
          description: Ошибка в запросе
        '500':
          description: Внутренняя ошибка сервера

  # Сервис доставки
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReserveWarehouseRequest:
      type: object
      properties:
        order_id:
          type: integer
        user_id:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
              quantity:
                type: integer
    ReleaseWarehouseRequest:
      type: object
      properties:
        order_id:
          type: integer
        user_id:
          type: integer
    ConfirmWarehouseRequest:
      type: object
      properties:
        order_id:
          type: integer
        user_id:
          type: integer 